폴더, 파일명 컨벤션에 맞게 수정
