+++ apps/app-front/src/libs/i18n/config.ts
