+++ apps/app-front/src/services/locale.ts
@@ -2,7 +2,7 @@
 
 import { cookies } from 'next/headers';
 
-import { Locale, defaultLocale } from '@/i18n/config';
+import { Locale, defaultLocale } from '@/libs/i18n/config';
 
 // In this example the locale is read from a cookie. You could alternatively
 // also read it from a database, backend service, or any other source.
