--- apps/back-office-front/src/config/index.ts
@@ -1,6 +1,0 @@
-export const env = {
-  // 클라이언트에서 사용 가능한 환경변수
-  publicEnv: {
-    apiUrl: process.env.NEXT_PUBLIC_API_URL,
-  },
-};
