+++ apps/back-office-front/src/configs/env.ts
@@ -1,0 +1,6 @@
+export const env = {
+  // 클라이언트에서 사용 가능한 환경변수
+  publicEnv: {
+    apiUrl: process.env.NEXT_PUBLIC_API_URL,
+  },
+} as const;
