+++ apps/ceo-front/src/app/layout.tsx
@@ -5,7 +5,7 @@ import { ReactNode, Suspense } from 'react';
 
 import Analytics from '@/components/GA/Analytics';
 import Layout from '@/components/Home/Layout';
-import { GA_TRACKING_ID } from '@/lib/GA/gtag';
+import { GA_TRACKING_ID } from '@/libs/ga/gtag';
 
 export default function RootLayout({ children }: { children: ReactNode }) {
   return (
