+++ apps/store-front/src/app/layout.tsx
@@ -4,7 +4,7 @@ import Script from 'next/script';
 import { ReactNode, Suspense } from 'react';
 
 import Analytics from '@/components/GA/Analytics';
-import { GA_TRACKING_ID } from '@/lib/GA/gtag';
+import { GA_TRACKING_ID } from '@/libs/ga/gtag';
 
 export default function RootLayout({ children }: { children: ReactNode }) {
   return (
