--- apps/store-front/src/components/Home/layout.tsx
@@ -1,13 +1,0 @@
-'use client';
-
-import { ReactNode } from 'react';
-
-interface LayoutProps {
-  children: ReactNode;
-}
-
-function Layout({ children }: LayoutProps) {
-  return <>{children}</>;
-}
-
-export default Layout;
