--- apps/store-front/src/components/Icon/CircleRightIcon.tsx
@@ -1,30 +1,0 @@
-import { ComponentPropsWithoutRef } from 'react';
-
-export interface IconProps extends ComponentPropsWithoutRef<'svg'> {
-  color?: string;
-  size?: number;
-  width?: number;
-  height?: number;
-}
-
-function CircleRightIcon({ size = 16, color = '#000000', ...props }: IconProps) {
-  return (
-    <svg
-      xmlns='http://www.w3.org/2000/svg'
-      width={size}
-      height={size}
-      viewBox='0 0 512 512'
-      fill='currentColor'
-      {...props}
-    >
-      <path
-        d='M464 256A208 208 0 1 1 48 256a208 208 0 1 1 416 0zM0 256a256 256 0 1 0 512 0A256 256 0 1 0 0 256zM294.6 151.2c-4.2-4.6-10.1-7.2-16.4-7.2C266 144 256 154 256 166.3l0 41.7-96 0c-17.7 0-32 14.3-32 32l0 32c0 17.7 14.3 32 32 32l96 0 0 41.7c0 12.3 10 22.3 22.3 22.3c6.2 0 12.1-2.6 16.4-7.2l84-91c3.5-3.8 5.4-8.7 5.4-13.9s-1.9-10.1-5.4-13.9l-84-91z'
-        stroke={color}
-        strokeWidth={1}
-        fillRule='evenodd'
-      />
-    </svg>
-  );
-}
-
-export default CircleRightIcon;
