--- apps/store-front/src/components/Icon/LeftArrowIcon.tsx
@@ -1,28 +1,0 @@
-import { ComponentPropsWithoutRef } from 'react';
-
-export interface IconProps extends ComponentPropsWithoutRef<'svg'> {
-  color?: string;
-  size?: number;
-  width?: number;
-  height?: number;
-}
-
-function LeftArrowIcon({ size = 16, color = '#000000', ...props }: IconProps) {
-  return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      width={size}
-      height={size}
-      viewBox="0 0 16 16"
-      fill="currentColor"
-      {...props}>
-      <path
-        d="M15 8a.5.5 0 0 0-.5-.5H2.707l3.147-3.146a.5.5 0 1 0-.708-.708l-4 4a.5.5 0 0 0 0 .708l4 4a.5.5 0 0 0 .708-.708L2.707 8.5H14.5A.5.5 0 0 0 15 8"
-        stroke={color}
-        fillRule="evenodd"
-      />
-    </svg>
-  );
-}
-
-export default LeftArrowIcon;
