+++ apps/store-front/src/configs/env.ts
@@ -1,0 +1,4 @@
+export const env = {
+  API_URL: process.env.NEXT_PUBLIC_API_URL,
+  GA_MEASUREMENT_ID: process.env.NEXT_PUBLIC_GA_MEASUREMENT_ID,
+} as const;
