--- apps/user-front/src/components/Home/layout.tsx
@@ -1,11 +1,0 @@
-import { ReactNode } from 'react';
-
-interface LayoutProps {
-  children: ReactNode;
-}
-
-function Layout({ children }: LayoutProps) {
-  return <>{children}</>;
-}
-
-export default Layout;
