--- apps/user-front/src/components/Icon/ChevronDownIcon.tsx
@@ -1,29 +1,0 @@
-import { ComponentPropsWithoutRef } from 'react';
-
-export interface IconProps extends ComponentPropsWithoutRef<'svg'> {
-  color?: string;
-  size?: number;
-  width?: number;
-  height?: number;
-}
-
-function ChevronDownIcon({ size = 16, color = '#000000', ...props }: IconProps) {
-  return (
-    <svg
-      xmlns='http://www.w3.org/2000/svg'
-      width={size}
-      height={size}
-      viewBox='0 0 512 512'
-      fill='currentColor'
-      {...props}
-    >
-      <path
-        d='M233.4 406.6c12.5 12.5 32.8 12.5 45.3 0l192-192c12.5-12.5 12.5-32.8 0-45.3s-32.8-12.5-45.3 0L256 338.7 86.6 169.4c-12.5-12.5-32.8-12.5-45.3 0s-12.5 32.8 0 45.3l192 192z'
-        stroke={color}
-        fillRule='evenodd'
-      />
-    </svg>
-  );
-}
-
-export default ChevronDownIcon;
