--- apps/user-front/src/components/Icon/RightArrowIcon.tsx
@@ -1,29 +1,0 @@
-import { ComponentPropsWithoutRef } from 'react';
-
-export interface IconProps extends ComponentPropsWithoutRef<'svg'> {
-  color?: string;
-  size?: number;
-  width?: number;
-  height?: number;
-}
-
-function RightArrowIcon({ width = 30, height = 16, color = '#000000', ...props }: IconProps) {
-  return (
-    <svg
-      xmlns='http://www.w3.org/2000/svg'
-      width={width}
-      height={height}
-      viewBox='0 0 512 512'
-      fill='currentColor'
-      {...props}
-    >
-      <path
-        d='M502.6 278.6c12.5-12.5 12.5-32.8 0-45.3l-128-128c-12.5-12.5-32.8-12.5-45.3 0s-12.5 32.8 0 45.3L402.7 224 32 224c-17.7 0-32 14.3-32 32s14.3 32 32 32l370.7 0-73.4 73.4c-12.5 12.5-12.5 32.8 0 45.3s32.8 12.5 45.3 0l128-128z'
-        stroke={color}
-        fillRule='evenodd'
-      />
-    </svg>
-  );
-}
-
-export default RightArrowIcon;
