--- apps/user-front/src/components/Icon/index.ts
@@ -1,7 +1,0 @@
-export { default as RightArrowIcon } from './RightArrowIcon';
-export { default as LeftArrowIcon } from './LeftArrowIcon';
-export { default as MenuBarIcon } from './MenuBarIcon';
-export { default as CheckIcon } from './CheckIcon';
-export { default as CloseIcon } from './CloseIcon';
-export { default as CircleRightIcon } from './CircleRightIcon';
-export { default as ChevronDownIcon } from './ChevronDownIcon';
