--- apps/user-front/src/config/index.ts
@@ -1,4 +1,0 @@
-export const config = {
-  API_URL: process.env.NEXT_PUBLIC_API_URL,
-  GA_MEASUREMENT_ID: process.env.NEXT_PUBLIC_GA_MEASUREMENT_ID,
-};
